backend/app/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.6.0
anthropic==0.8.1
anyio==3.7.1
anytree==2.12.1
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.29.4
azure-storage-blob==12.18.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.2
camel-converter==3.1.1
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
coreschema==0.0.4
cryptography==41.0.1
dataclasses-json==0.5.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
distro==1.8.0
Django==4.2.5
django-compat==1.0.15
django-cors-headers==4.4.0
django-filter==2.4.0
django-mssql-backend==2.8.1
django-ordered-model==3.4.3
django-sendgrid-v5==1.2.0
django-storages==1.14
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
dnspython==2.4.2
ecdsa==0.18.0
et-xmlfile==1.1.0
executing==2.0.1
fairscale==0.4.13
faiss-cpu==1.7.4
filelock==3.8.2
fire==0.4.0
Flask==3.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
fsspec==2023.12.2
gitdb==4.0.11
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-status==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.20.2
hyperframe==6.0.1
idna==3.4
intel-openmp==2021.4.0
ipykernel==6.29.0
ipython==8.20.0
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.1
lancedb==0.4.0
langchain==0.0.352
langchain-community==0.0.13
langchain-core==0.1.12
langsmith==0.0.83
lazy-object-proxy==1.10.0
libcst==1.0.1
loguru==0.6.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
meilisearch==0.21.0
# Editable install with no version control (metagpt==0.6.0)
-e c:\users\divyo\appdata\local\programs\python\python311\lib\site-packages\metagpt-0.6.0-py3.11.egg
mkl==2021.4.0
more-itertools==10.2.0
motor==3.3.2
mpmath==1.3.0
mssql-django==1.3
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==1.6.0
openapi-core==0.18.2
openapi-schema-pydantic==1.2.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.0.3
parse==1.20.0
parso==0.8.3
pathable==0.4.3
Pillow==9.5.0
platformdirs==2.6.0
portalocker==2.8.2
prance==23.6.21.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
py==1.11.0
pyarrow==14.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.0
pymongo==4.6.1
pyodbc==4.0.39
pyspnego==0.9.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
python-pptx==0.6.21
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.7.0
ratelimiter==1.2.0.post0
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
semantic-kernel==0.4.3.dev0
semver==3.0.2
sendgrid==6.10.0
sentencepiece==0.2.0
SharePlum==0.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.3
stack-data==0.6.3
starkbank-ecdsa==2.2.0
sympy==1.12
ta==0.10.2
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.2
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.3.0
tornado==6.4
tqdm==4.65.0
traitlets==5.14.1
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.17.1
wcwidth==0.2.13
websocket-client==1.6.2
websockets==12.0
Werkzeug==3.0.0
win32-setctime==1.1.0
wrapt==1.15.0
XlsxWriter==3.1.2
yarl==1.9.2
zhipuai==2.0.1
